GH Actions: disallow for Symfony/Flex composer plugin
As discussed in the ticket, as of July 1st 2022, Composer requires explicit permission to be given for Composer plugins to be run when using Composer 2.2+.
See: https://blog.packagist.com/composer-2-2/#more-secure-plugin-execution

On top of that the website which Symfony Flex tries to access during its install is not responding anymore when using old Flex versions in which those API calls haven't been updated yet.
Also see: https://symfony.com/blog/upgrade-flex-on-your-symfony-projects

Combining the above two issues and having done some test runs on GH Actions, I've come to the conclusion that not allowing the `symfony/flex` plugin and disabling the running of plugins during the install step solves the problems without any impact on the tests.

So that is what this commit now does.