Composer: limit allowed PHPUnit 9.x version
The PHPUnit 8.5.29/9.5.23 patch update removed the dependency on Prophecy, which means that projects using Prophecy now need to require it themselves:

> sebastianbergmann/phpunit 5033: Do not depend on phpspec/prophecy

See: https://github.com/sebastianbergmann/phpunit/blob/main/ChangeLog-9.5.md

This, however, has its own problems, what with the `phpspec/propephy-phpunit` layer not allowing installation with PHPUnit < 9.

So, for now, this limits PHPUnit to a version before the dependency was dropped.