README: fix broken badge and mention GH Actions
The build status badge in README was broken as it still pointed to Travis, while CI has been moved to GH Actions.

This commit fixes this by replacing the Travis badge with two badges for the GH Actions workflows.

And while GH Actions is already mentioned lower down in the README with a code example of how to use PHP Coveralls, it wasn't mentioned in the "Prerequisites" at the top of the README.
Fixed now.